// src/types/employees-types.ts
export interface Employee {
  EmployeeInfoID: number;
  SN: number;
  Fullname: string;
  Address: string;
  Phone: string;
  Email: string;
  DOB: string;
  DepartmentID: number;
  DepartmentName: string;
  BranchID: number;
  BranchName: string;
  MainBranchID: number;
  MainBranchName: string;
  Gender: number;
  EmpStatus: number;
  Status: number;
  OrganizationOfficeID: number;
  Photo: string;
}